Rename the translation model to locale
